2nd & subsequent endpoints requires unique ports
`hasura console` for additional running instances requires we manually specify the endpoint ***and*** the console and api ports as different (e.g. 9700, 9703) than the default (9695, 9693).  Otherwise, `hasura console` will see the 1st instance ports in use and simply connect to that instance, *regardless* of `--endpoint <endpoint>` being specified & pointing to 2nd instance.  Ports will not be auto-negotiated by `hasura console` when a 2nd instance is running.