Add haddocks index file
Because we create a new version of the haddocks for every release, the
available haddocks index will keep growing, and we'll keep having to
update the `gh-pages` markdown files. To avoid that, we add a haddocks
index file that uses the GitHub Contents API to find the available
haddocks, and list them.