console: Factor out downloading assets.
This pulls the script to download the common assets out of the
`server-build` Makefile target into its own target. Because we use the
output directory name as the target, we allow `make` to decide whether
it's necessary.

As there are no inputs to the target, it will only be done once.

Behavior should be unaffected.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/4617
GitOrigin-RevId: 15d64467c92be761f2738d179b1b027abe8bd8ca